Simplification du code. Meilleur utilisation du type générique TIn.

+++ GameMath.Demo/Demo.cs
@@ -16,7 +16,7 @@ namespace GameMath.Demo
         private int _bounceX;
         private Point2D<int> _backPoint;
         private ITween<int, decimal> _elasticTween;
-        private ITween<int, decimal> _bounceTween;
+        private ITween<float, decimal> _bounceTween;
         private ITween<int, Point2D<decimal>> _backTween;
 
         public Demo()
@@ -26,7 +26,7 @@ namespace GameMath.Demo
             IsMouseVisible = true;
 
             _elasticTween = TweenBuilder.From(100).To(600).For(2000f).Elastic().EaseOut().LoopReverse().Build();
-            _bounceTween = TweenBuilder.From(100).To(600).For(2000f).Bounce().EaseOut().Build();
+            _bounceTween = TweenBuilder.From(100f).To(600f).For(2000f).Bounce().EaseOut().Build();
             _backTween = TweenBuilder.From(new Microsoft.Xna.Framework.Point() { X = 100, Y = 400 })
                 .To(new Microsoft.Xna.Framework.Point() { X = 600, Y = 400 }).For(1500f).Back(1.5).EaseInOut().Build();
         }
